requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
pandas
requests
openai
google-analytics-data
httpx
